Initial commit for UI Testing of ContactListApp
- Started working on POM
- inserted logic for work with Drivers / Waits
- Created BasePage / LoginPage / UserPage
- put some 'TODOs' for future work, so I know where I've left off
