+++ src/main/java/com/ContactList/API/core/payloads/UserPayloads/UserLoginPayload.java
@@ -1,0 +1,12 @@
+package com.ContactList.API.core.payloads.UserPayloads;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class UserLoginPayload {
+
+    private String email;
+    private String password;
+}
