+++ src/main/java/com/ContactList/API/enums/endpoints/ContactEndpoints.java
@@ -1,0 +1,18 @@
+package com.ContactList.API.enums.endpoints;
+
+public enum ContactEndpoints {
+
+    ME("me"),
+    LOGIN("login"),
+    LOGOUT("logout");
+
+    private final String endpoint;
+
+    ContactEndpoints(String endpoint) {
+        this.endpoint = endpoint;
+    }
+
+    public String getEndpoint() {
+        return endpoint;
+    }
+}
