+++ src/main/java/com/ContactList/API/enums/endpoints/UserEndpoints.java
@@ -1,0 +1,18 @@
+package com.ContactList.API.enums.endpoints;
+
+public enum UserEndpoints {
+
+    ME("me"),
+    LOGIN("login"),
+    LOGOUT("logout");
+
+    private final String endpoint;
+
+    UserEndpoints(String endpoint) {
+        this.endpoint = endpoint;
+    }
+
+    public String getEndpoint() {
+        return endpoint;
+    }
+}
