+++ src/main/java/com/ContactList/API/enums/headers/Headers.java
@@ -1,0 +1,16 @@
+package com.ContactList.API.enums.headers;
+
+import lombok.Getter;
+
+@Getter
+public enum Headers {
+
+    AUTHORIZATION("Authorization");
+
+    private final String header;
+
+    Headers(String header) {
+        this.header = header;
+    }
+
+}
