+++ src/main/java/com/ContactList/API/utils/dataManagement/DataSerializer.java
@@ -1,0 +1,17 @@
+package com.ContactList.API.utils.dataManagement;
+
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public class DataSerializer {
+
+    private static final ObjectMapper mapper = new ObjectMapper();
+
+    public static String serializeUserPayload(Object payload) {
+        try {
+            return mapper.writeValueAsString(payload);
+        } catch (Exception e) {
+            throw new RuntimeException("failed to serialize an object");
+        }
+    }
+}
