+++ src/main/java/com/ContactList/UI/enums/AppEndpoints.java
@@ -1,0 +1,16 @@
+package com.ContactList.UI.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum AppEndpoints {
+
+    BASE_PATH("https://thinking-tester-contact-list.herokuapp.com/");
+
+    private String endpoint;
+
+    private AppEndpoints(String endpoint) {
+        this.endpoint = endpoint;
+    }
+
+}
