+++ src/main/java/com/ContactList/UI/pages/UserPage.java
@@ -1,0 +1,9 @@
+package com.ContactList.UI.pages;
+
+//TODO: TBD
+public class UserPage extends BasePage {
+
+    public UserPage() {
+        super();
+    }
+}
