--- src/main/java/com/ContactList/config/enums/endpoints/UserEndpoints.java
@@ -1,18 +1,0 @@
-package com.ContactList.config.enums.endpoints;
-
-public enum UserEndpoints {
-
-    ME("me"),
-    LOGIN("login"),
-    LOGOUT("logout");
-
-    private final String endpoint;
-
-    UserEndpoints(String endpoint) {
-        this.endpoint = endpoint;
-    }
-
-    public String getEndpoint() {
-        return endpoint;
-    }
-}
