--- src/main/java/com/ContactList/config/enums/headers/Headers.java
@@ -1,16 +1,0 @@
-package com.ContactList.config.enums.headers;
-
-import lombok.Getter;
-
-@Getter
-public enum Headers {
-
-    AUTHORIZATION("Authorization");
-
-    private final String header;
-
-    Headers(String header) {
-        this.header = header;
-    }
-
-}
