--- src/main/java/com/ContactList/core/payloads/UserPayloads/UserLoginPayload.java
@@ -1,12 +1,0 @@
-package com.ContactList.core.payloads.UserPayloads;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class UserLoginPayload {
-
-    private String email;
-    private String password;
-}
