+++ src/test/java/com/ContactList/tests/api/UserTests.java
@@ -1,10 +1,10 @@
 package com.ContactList.tests.api;
 
-import com.ContactList.core.payloads.UserPayloads.UserBodyPayload;
-import com.ContactList.core.responses.userResponses.UserResponse;
-import com.ContactList.core.services.UserService;
-import com.ContactList.utils.dataManagement.DataGenerator;
-import com.ContactList.utils.helpers.UserApiHelper;
+import com.ContactList.API.core.payloads.UserPayloads.UserBodyPayload;
+import com.ContactList.API.core.responses.userResponses.UserResponse;
+import com.ContactList.API.core.services.UserService;
+import com.ContactList.API.utils.dataManagement.DataGenerator;
+import com.ContactList.API.utils.helpers.UserApiHelper;
 import io.qameta.allure.Feature;
 import io.qameta.allure.Owner;
 import io.qameta.allure.Story;
